pyproject.toml
[project]
name = "face-analyzer"
version = "1.0.0"

requires-python = "==3.10"
dependencies = [
    "common-code[test] @ git+https://github.com/swiss-ai-center/common-code.git@main",
    "deepface==0.0.89",
    "numpy==1.26.4",
    "pillow==10.2.0",
    "tf-keras==2.16.0"

]
[tool.pytest.ini_options]
pythonpath = [".", "src"]
addopts = "--cov-config=.coveragerc --cov-report xml --cov-report term-missing --cov=./src"

[dependency-groups]
dev = [
    "black>=26.3.1",
    "ruff>=0.15.7",
]
